Committing a test of the previous version of _printf_binary

+++ _printf.c
@@ -69,9 +69,9 @@ int _printf_string(char *tmp)
  * @count: pointer
  * Return: int
  */
-int _printf_binary(unsigned int num, int *count)
+int _printf_binary(unsigned long int num, int *count)
 {
-	unsigned int binary = 0, base = 1;
+	unsigned long int binary = 0, base = 1;
 
 	if (num == 0)
 		(*count)++;
@@ -117,7 +117,7 @@ int _printf_helper(const char *format, va_list arg, int *i)
 		_rec_number(num, &res);
 		break;
 	case 'b':
-		num = va_arg(arg, unsigned int);
+		num = va_arg(arg, unsigned long int);
 		_printf_binary(num, &res);
 		break;
 	default:
